refactor(core): move tool manifest owner into product runtime
